ENH: add support for mash distance in distance app
